3.4-django-testing/django_testing/requirements.txt
django~=4.1.7
djangorestframework~=3.14.0
django-filter~=23.1
psycopg2-binary

pytest~=7.2.2